Feat: 상세페이지 관련 컴포넌트 경로 변경 및 반응형, 스타일 구현
