+++ src/components/domains/detail/InfoAndReviewTab/InfoAndReviewTab.module.scss
@@ -5,6 +5,11 @@
   border-bottom: solid 0.1rem $gray20;
   padding-bottom: 2rem;
 
+  @include responsive("M") {
+    font-size: 1.4rem;
+    padding-bottom: 1.5rem;
+  }
+
   &-btn {
     cursor: pointer;
