+++ src/components/domains/detail/course/CourseInfo/TravelCourseOnMap/DestinationCard.module.scss
@@ -1,13 +1,16 @@
 .container {
+  @include flexbox(start, center);
   background-color: $white;
   padding: 1.2rem 2rem;
   border-radius: 1rem;
   box-shadow: 0px 3px 16px 0px rgba(0, 0, 0, 0.08);
+  width: 30rem;
   cursor: pointer;
 
   .content {
     @include flexbox(stretch, center);
     gap: 2rem;
+    width: 100%;
 
     .icon {
       width: 3.5rem;
@@ -17,18 +20,31 @@
       &-num {
         @include text-style(1.2, 700, $black30);
         @include pos-center-x();
-        top: 1rem;
+        top: 1.05rem;
       }
     }
 
     .text {
+      width: 100%;
+      overflow: hidden;
+
       &-title {
         @include text-style(1.4, 700, $black30);
+        width: 100%;
         padding-bottom: 0.5rem;
+        overflow: hidden;
+        white-space: nowrap;
+        text-overflow: ellipsis;
+        word-break: break-all;
       }
 
       &-address {
+        width: 100%;
         @include text-style(1.2, 500, $black30);
+        overflow: hidden;
+        white-space: nowrap;
+        text-overflow: ellipsis;
+        word-break: break-all;
       }
     }
   }
