+++ src/components/domains/detail/course/CourseInfo/TravelCourseOnMap/DestinationCard.tsx
