+++ src/components/domains/detail/course/CourseInfo/TravelDurationAndGroup/TravelDurationAndGroup.tsx
@@ -1,25 +1,16 @@
-import { getCourseDetail } from "@/api/course";
 import Image from "@/components/commons/Image";
-import { defaultCourseDetail } from "@/constants/defaultValues";
 import { IMAGES } from "@/constants/images";
-import { useQuery } from "@tanstack/react-query";
 import classNames from "classnames/bind";
-import { useParams } from "react-router-dom";
 import styles from "./TravelDurationAndGroup.module.scss";
 
 const cx = classNames.bind(styles);
 
-const TravelDurationAndGroup = () => {
-  const { id } = useParams<{ id: string }>();
+interface TravelDurationAndGroupProps {
+  travelerCount: number;
+  duration: number;
+}
 
-  const { data: courseDetailData } = useQuery({
-    queryKey: ["courseDetailData", id],
-    queryFn: () => getCourseDetail(Number(id)),
-    retry: 0,
-  });
-
-  const courseDetail = courseDetailData ?? defaultCourseDetail;
-
+const TravelDurationAndGroup = ({ travelerCount, duration }: TravelDurationAndGroupProps) => {
   return (
     <ul className={cx("container")}>
       <li className={cx("element")}>
@@ -28,7 +19,7 @@ const TravelDurationAndGroup = () => {
         </div>
         <div className={cx("text")}>
           <h4>여행추천인원</h4>
-          <p className={cx("text-stress")}>{courseDetail.travelerCount}인</p>
+          <p className={cx("text-stress")}>{travelerCount}인</p>
         </div>
       </li>
       <li className={cx("element")}>
@@ -37,7 +28,7 @@ const TravelDurationAndGroup = () => {
         </div>
         <div className={cx("text")}>
           <h4>여행기간</h4>
-          <p className={cx("text-stress")}>{courseDetail.duration}일</p>
+          <p className={cx("text-stress")}>{duration}일</p>
         </div>
       </li>
     </ul>
