+++ src/layout/DetailPageLayout/DetailPageLayout.module.scss
@@ -11,4 +11,9 @@
       padding: 0 1rem;
     }
   }
+
+  .article {
+    padding: 5rem 0 10rem;
+    width: 100%;
+  }
 }
