+++ src/pages/CourseDetailPage/index.tsx
@@ -1,12 +1,10 @@
 import CourseHeader from "@/components/domains/detail/course/CourseHeader/CourseHeader";
+import CourseInfo from "@/components/domains/detail/course/CourseInfo/CourseInfo";
 import CourseMain from "@/components/domains/detail/course/CourseMain/CourseMain";
-import TravelCourseOnMap from "@/components/domains/detail/course/TravelCourseOnMap";
 import DetailPageLayout from "@/layout/DetailPageLayout";
 
 const CourseDetailPage = () => {
-  return (
-    <DetailPageLayout type="course" header={<CourseHeader />} main={<CourseMain />} info={<TravelCourseOnMap />} />
-  );
+  return <DetailPageLayout type="course" header={<CourseHeader />} main={<CourseMain />} info={<CourseInfo />} />;
 };
 
 export default CourseDetailPage;
